# tests/security/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TEE Vote Backend API",
    "version": "1.0.0",
    "description": "OpenAPI contract used by OWASP ZAP API Scan in the security test suite."
  },
  "servers": [
    {
      "url": "http://backend:3001"
    }
  ],
...
  "paths": {
...
    "/api/mis-postulaciones/{formId}": {
      "get": {
        "tags": [
          "MyApplications"
        ],
        "operationId": "getMyApplication",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "1f2e3d4c-5b6a-4978-8899-aabbccddeeff"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonObject"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "tags": [
          "MyApplications"
        ],
        "operationId": "saveMyApplication",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "1f2e3d4c-5b6a-4978-8899-aabbccddeeff"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "last_name_1": {
                    "type": "string",
                    "nullable": true
                  },
                  "last_name_2": {
                    "type": "string",
                    "nullable": true
                  },
                  "first_name": {
                    "type": "string",
                    "nullable": true
                  },
                  "national_id": {
                    "type": "string",
                    "nullable": true,
                    "example": "119330447"
                  },
                  "carnet": {
                    "type": "string",
                    "nullable": true,
                    "example": "2024080534"
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true,
                    "example": "88887777"
                  },
                  "sede": {
                    "type": "string",
                    "nullable": true
                  },
                  "career": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonObject"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/mis-postulaciones/{formId}/enviar": {
      "post": {
        "tags": [
          "MyApplications"
        ],
        "operationId": "submitMyApplication",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "1f2e3d4c-5b6a-4978-8899-aabbccddeeff"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonObject"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "JsonObject": {
        "description": "JSON object response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "error": {
            "type": "string",
            "example": "No autenticado"
          },
          "message": {
            "type": "string",
            "example": "Solicitud invalida"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          }
        }
      },
...
    }
  }
}